Implementing some improvements
-Deleted #pragma once in order to decrease project size
-Improved logic of state classes (added ReCompile method, updated other methods)

+++ BrainfuckCompiler/Storage.h
@@ -1,5 +1,5 @@
 #pragma once
 struct Storage {
     char bfstorage[30000] = {};
-    size_t bfindex = 0;
+    short bfindex = 0;
 };
